+++ src/tests/fakes/items.c
@@ -21,7 +21,6 @@
 #include <trx/game/objects/families.h>
 #include <trx/game/objects/names.h>
 #include <trx/game/objects/property.h>
-#include <trx/game/objects/vars.h>
 #include <trx/game/pathing/lot.h>
 #include <trx/game/rooms.h>